:truck: Move set-language-cookie util to translations app
openforms/utils/translation is more aimed at the gettext level
rather than runtime utilities.
