+++ src/openforms/utils/translations.py
@@ -4,8 +4,6 @@ from typing import Callable
 from django.conf import settings
 from django.utils import translation
 
-from rest_framework.response import Response
-
 
 def get_default(value) -> str:
     return str(value)
@@ -46,16 +44,3 @@ class ensure_default_language(translation.override):
         if current is not None:
             self.language = current
         return super().__enter__()
-
-
-def set_language_cookie(response: Response, language_code: str) -> None:
-    response.set_cookie(
-        key=settings.LANGUAGE_COOKIE_NAME,
-        value=language_code,
-        max_age=settings.LANGUAGE_COOKIE_AGE,
-        domain=settings.LANGUAGE_COOKIE_DOMAIN,
-        httponly=settings.LANGUAGE_COOKIE_HTTPONLY,
-        path=settings.LANGUAGE_COOKIE_PATH,
-        samesite=settings.LANGUAGE_COOKIE_SAMESITE,
-        secure=settings.LANGUAGE_COOKIE_SECURE,
-    )
